Code cleanup: moving chunks into separate modules
